Fix detection of whether `long double` is the same as `double`
One change in PR 305 was not quite right: it changed a check for whether
`$(ARCH)` is `aarch64` to a check for whether it's `aarch64 riscv64`,
which can never be true. It was a non-functional change on RISC-V (since
it can't be true) but broke AArch64, at least for FreeBSD. Notably,
changing the check to what I assume was intended, i.e. `aarch64` or
`riscv64`, breaks RISC-V. So this commit instead reverts this one change
in 305.